refactor: extract TabCreateNFT and TabHome functions into separate files
